refactor(aleo): move ContractCall to aleo-gmp-types
1. Move `ContractCall` struct and its trait implementations from `ampd/src/aleo/generated_structs.rs` to `packages/aleo-gmp-types/src/aleo_struct/generated_structs.rs`.
2. Rename function `find_in_outputs` to `find_signers_rotated_in_outputs` in `ampd/src/aleo/utils.rs` and update all usages.
3. Apply `cargo fmt`.
